решение замечаний ментора
