--- src/main/java/userRegistration/Main.java
@@ -1,8 +1,0 @@
-package userRegistration;
-
-public class Main {
-    public static void main(String[] args) {
-        User user = new User("Neymar", 30, "Santos FC", " some Street in Brazil");
-
-    }
-}
